move xls upload to util
