+++ assistant/views/mandate.py
@@ -92,174 +92,3 @@ def load_mandates(request):
     return layout.render(request, "load_mandates.html", {})
 
 
-@user_passes_test(user_is_manager, login_url='assistants_home')
-def upload_mandates_file(request):
-    global ASSISTANTS_IMPORTED, ASSISTANTS_UPDATED, MANDATES_IMPORTED, MANDATES_UPDATED, PERSONS_NOT_FOUND
-    ASSISTANTS_UPDATED = 0
-    ASSISTANTS_IMPORTED = 0
-    MANDATES_IMPORTED = 0
-    MANDATES_UPDATED = 0
-    PERSONS_NOT_FOUND = 0
-    if request.method == 'POST':
-        form = MandateFileForm(request.POST, request.FILES)
-        if form.is_valid():
-            file_name = request.FILES['file']
-            if file_name is not None:
-                try:
-                    read_xls_mandates(request, file_name)
-                except IndexError:
-                    messages.add_message(request, messages.ERROR,
-                                         'xls_columns_structure_error'.format('via_excel', 'get_excel_file'))
-        else:
-            for error_msg in [error_msg for error_msgs in form.errors.values() for error_msg in error_msgs]:
-                messages.add_message(request, messages.ERROR, "{}".format(error_msg))
-        return layout.render(request, "load_mandates.html", {'imported_assistants': ASSISTANTS_IMPORTED,
-                                                             'imported_mandates': MANDATES_IMPORTED,
-                                                             'updated_mandates': MANDATES_UPDATED,
-                                                             'updated_assistants': ASSISTANTS_UPDATED,
-                                                             'persons_not_found': PERSONS_NOT_FOUND})
-
-
-@user_passes_test(user_is_manager, login_url='assistants_home')
-def read_xls_mandates(request, file_name):
-    try:
-        workbook = load_workbook(file_name, read_only=True, data_only=True)
-    except KeyError:
-        messages.add_message(request, messages.ERROR, 'file_must_be_xlsx')
-        return False
-    first_sheet = workbook.get_sheet_names()[0]
-    worksheet = workbook.get_sheet_by_name(first_sheet)
-    titles_row = []
-    current_row = 1
-    for row in worksheet.iter_rows():
-        if current_row == 1:
-            titles_row = save_xls_rows_titles(row)
-            if check_file_format(request, titles_row) is False:
-                return False
-        else:
-            current_record = xls_row_to_dict(row, titles_row)
-            assistant = create_academic_assistant_if_not_exists(current_record)
-            if assistant:
-                mandate = create_assistant_mandate_if_not_exists(current_record, assistant)
-                institute = search_structure_by_acronym_and_type(current_record.get('INSTITUTE'), 'INSTITUTE')
-                if institute:
-                    link_mandate_to_structure(mandate, institute)
-                faculty = search_structure_by_acronym_and_type(current_record.get('FACULTY'), 'FACULTY')
-                if faculty:
-                    link_mandate_to_structure(mandate, faculty)
-                pole = search_structure_by_acronym_and_type(current_record.get('POLE'), 'POLE')
-                if pole:
-                    link_mandate_to_structure(mandate, pole)
-                program_commission = search_structure_by_acronym_and_type(current_record.get('PROGRAM_COMMISSION'),
-                                                                          'PROGRAM_COMMISSION')
-                if program_commission:
-                    link_mandate_to_structure(mandate, program_commission)
-        current_row += 1
-
-
-def search_structure_by_acronym_and_type(acronym, type):
-    if not acronym:
-        return None
-    results = mdl.structure.search(acronym=acronym, type=type)
-    if len(results) > 0:
-        return results[0]
-    else:
-        return None
-
-
-def save_xls_rows_titles(current_row):
-    titles = []
-    for cell in current_row:
-        titles.append(cell.value)
-    return titles
-
-
-def xls_row_to_dict(row, titles):
-    record_to_import = {}
-    current_col = 0
-    for cell in row:
-        record_to_import[titles[current_col]] = cell.value
-        current_col += 1
-    return record_to_import
-
-
-def create_academic_assistant_if_not_exists(record):
-    global ASSISTANTS_UPDATED, ASSISTANTS_IMPORTED, PERSONS_NOT_FOUND
-    person = mdl.person.find_by_global_id(record.get('FGS'))
-    if person:
-        try:
-            assistant = assistant_mdl.academic_assistant.find_by_person(mdl.person.find_by_global_id(record.get('FGS')))
-            ASSISTANTS_UPDATED += 1
-        except ObjectDoesNotExist:
-            assistant = assistant_mdl.academic_assistant.AcademicAssistant()
-            ASSISTANTS_IMPORTED += 1
-            assistant.person = person
-            assistant.save()
-        return assistant
-    else:
-        PERSONS_NOT_FOUND += 1
-        return None
-
-
-def create_assistant_mandate_if_not_exists(record, assistant):
-    global MANDATES_IMPORTED, MANDATES_UPDATED
-    current_academic_year = mdl.academic_year.current_academic_year()
-    mandates = assistant_mdl.assistant_mandate.find_mandate(assistant, current_academic_year, record.get('SAP_ID'))
-    if len(mandates) == 0:
-        mandate = assistant_mdl.assistant_mandate.AssistantMandate()
-        MANDATES_IMPORTED += 1
-    else:
-        mandate = mandates[0]
-        MANDATES_UPDATED += 1
-    mandate.assistant = assistant
-    mandate.academic_year = current_academic_year
-    mandate.end_date = record.get('END_DATE')
-    mandate.entry_date = record.get('ENTRY_DATE')
-    mandate.fulltime_equivalent = record.get('FULLTIME_EQUIVALENT')
-    mandate.sap_id = record.get('SAP_ID')
-    mandate.contract_duration = record.get('CONTRACT_DURATION')
-    mandate.contract_duration_fte = record.get('CONTRACT_DURATION_FTE')
-    mandate.renewal_type = record.get('RENEWAL_TYPE')
-    mandate.absences = record.get('ABSENCES')
-    mandate.comment = record.get('COMMENT')
-    mandate.other_status = record.get('OTHER_STATUS')
-    mandate.assistant_type = record.get('ASSISTANT_TYPE_CODE')
-    mandate.grade = record.get('GRADE')
-    mandate.scale = record.get('SCALE')
-    mandate.state = 'TO_DO'
-    mandate.save()
-    return mandate
-
-
-def link_mandate_to_structure(mandate, structure=None):
-    if structure != 'None':
-        mandate_structures = assistant_mdl.mandate_structure.find_by_mandate_and_structure(mandate, structure)
-        if len(mandate_structures) == 0:
-            mandate_structure = assistant_mdl.mandate_structure.MandateStructure()
-        else:
-            mandate_structure = mandate_structures[0]
-        mandate_structure.assistant_mandate = mandate
-        mandate_structure.structure = structure
-        mandate_structure.save()
-        return mandate_structure
-    else:
-        return None
-
-
-def show_import_result(request):
-    global ASSISTANTS_IMPORTED, ASSISTANTS_UPDATED, MANDATES_IMPORTED, MANDATES_UPDATED, PERSONS_NOT_FOUND
-    return layout.render(request, "load_mandates.html", {'imported_assistants': ASSISTANTS_IMPORTED,
-                                                         'imported_mandates': MANDATES_IMPORTED,
-                                                         'updated_mandates': MANDATES_UPDATED,
-                                                         'updated_assistants': ASSISTANTS_UPDATED,
-                                                         'persons_not_found': PERSONS_NOT_FOUND})
-
-
-def check_file_format(request, titles_rows):
-    if len(titles_rows) != COLS_NUMBER:
-        messages.add_message(request, messages.ERROR, _('columns_number_error'))
-        return False
-    if titles_rows != COLS_TITLES:
-        messages.add_message(request, messages.ERROR, _('columns_title_error'))
-        messages.add_message(request, messages.ERROR, COLS_TITLES)
-        return False
